==> 2_MLS_Neighborhood_Data_ETL.ipynb <==
import pandas as pd
import numpy as np
import os as os

# SQL Alchemy
from sqlalchemy import create_engine

# Create Engine for employee db
engine = create_engine('postgresql://localhost:5432/Minneapolis_Police_Force_db')
connection = engine.connect()

#load csv into dataframe
load_csv = "source_files/MLPS_Neighborhoods_raw.csv"
MLPS_Neighborhoods_df = pd.read_csv(load_csv)
MLPS_Neighborhoods_df.head()

#Fetch community data from Postgres
community_df = pd.read_sql("SELECT community_id, name FROM community", connection)
                              
community_df.head()   

#Join neighborhood data and community df to get community id for each neighborhood
neighborhood_community=pd.merge(MLPS_Neighborhoods_df, community_df, how='inner', left_on='COMMUNITY', right_on='name')

neighborhood_community.head()

#Count number of neighborhoods in both dfs to validate if counts match
print(MLPS_Neighborhoods_df['NEIGHBORHOOD'].count())
print(neighborhood_community['NEIGHBORHOOD'].count())

#Drop fields from df which are not required in final output
neighborhood_community=neighborhood_community[['NEIGHBORHOOD','community_id']]
neighborhood_community.head()

#Rename df names to match Postgres table
neighborhood_community.rename(columns={'NEIGHBORHOOD': 'name'}, inplace=True)
neighborhood_community.head()

#Start index from 1 in MLPS_Communities_df and rename index field
neighborhood_community.index = np.arange(1, len(neighborhood_community)+1)
neighborhood_community.index = neighborhood_community.index.set_names(['neighborhood_id'])

neighborhood_community.head()

#append data to Postgres existing table
neighborhood_community.to_sql('neighborhood', engine,if_exists='append')

#Save data to a csv file
output_data_file = "target_files/MLS_Neighborhoods.csv"
output_path = os.path.join(output_data_file)
neighborhood_community.to_csv(output_path, header=True)
