==> 4b_MLS_Income_by_Community_Transform_Load.ipynb <==
import pandas as pd
import numpy as np
import os as os

# SQL Alchemy
from sqlalchemy import create_engine

# Create Engine for employee db
engine = create_engine('postgresql://localhost:5432/Minneapolis_Police_Force_db')
connection = engine.connect()

#load csv into dataframe
load_csv = "resources/MLPS_Hsld_Income_by_Community.csv"
MLPS_Community_Income_df = pd.read_csv(load_csv)
MLPS_Community_Income_df.head()

# Transform Household Income Data for MLS communities

#Replace \n, commas, %, $ in source data with ''
MLPS_Community_Income_df = MLPS_Community_Income_df.replace('\n','', regex=True)
MLPS_Community_Income_df = MLPS_Community_Income_df.replace(',','', regex=True)
MLPS_Community_Income_df = MLPS_Community_Income_df.replace('%','', regex=True)
MLPS_Community_Income_df = MLPS_Community_Income_df.replace('$','', regex=True)
MLPS_Community_Income_df.head()

#Trim whitespace from ends of each value across all series in dataframe

MLPS_Community_Income_df = MLPS_Community_Income_df.applymap(lambda x: x.strip() if isinstance(x, str) else x)
MLPS_Community_Income_df.head()

#Extract the median income by removing '$ ' and converting to float

MLPS_Community_Income_df['Median_Income_Total'] = MLPS_Community_Income_df['Median_Income_Total'].str[2:].astype(float)
MLPS_Community_Income_df.head()

#Replace 'suppressed' and blank fields with NaN
MLPS_Community_Income_df= MLPS_Community_Income_df.replace('suppressed', np.nan)
MLPS_Community_Income_df= MLPS_Community_Income_df.replace('', np.nan)

#Covert all count fields to numeric and if there are invalid numbers

MLPS_Community_Income_df['IncomeLess35000_count'] = pd.to_numeric(MLPS_Community_Income_df['IncomeLess35000_count'])

MLPS_Community_Income_df['IncomeLess35to49_count'] = pd.to_numeric(MLPS_Community_Income_df['IncomeLess35to49_count'])

MLPS_Community_Income_df['IncomeLess50to74_count'] = pd.to_numeric(MLPS_Community_Income_df['IncomeLess50to74_count'])

MLPS_Community_Income_df['IncomeLess75to99_count'] = pd.to_numeric(MLPS_Community_Income_df['IncomeLess75to99_count'])

MLPS_Community_Income_df['Income100Plus_count'] = pd.to_numeric(MLPS_Community_Income_df['Income100Plus_count'])

MLPS_Community_Income_df.head()

#Replace '%' in percent field and convert to float

MLPS_Community_Income_df['IncomeLess35000_percent'] = MLPS_Community_Income_df['IncomeLess35000_percent'].astype(float)

MLPS_Community_Income_df['IncomeLess35to49_percent'] = MLPS_Community_Income_df['IncomeLess35to49_percent'].astype(float)

MLPS_Community_Income_df['IncomeLess50to74_percent'] = MLPS_Community_Income_df['IncomeLess50to74_percent'].astype(float)

MLPS_Community_Income_df['IncomeLess75to99_percent'] = MLPS_Community_Income_df['IncomeLess75to99_percent'].astype(float)

MLPS_Community_Income_df['Income100Plus_percent'] = MLPS_Community_Income_df['Income100Plus_percent'].astype(float)

MLPS_Community_Income_df.head()

#Count of records before join
MLPS_Community_Income_df['Community'].count()

#Fetch neighborhood data from Postgres
community_df = pd.read_sql("SELECT community_id, name FROM community", connection)
                              
community_df.head() 

#Join neighborhood data and MLPS_Community_Income_df df to get community id for each neighborhood
community_income=pd.merge(MLPS_Community_Income_df, community_df, how='inner', left_on='Community', right_on='name')

community_income.head()

#Count of records after join
community_income['Community'].count()

#Extract columns to be used in final target
clean_df = community_income[['community_id','IncomeLess35000_count','IncomeLess35000_percent','IncomeLess35to49_count',
                                'IncomeLess35to49_percent','IncomeLess50to74_count','IncomeLess50to74_percent',
                                'IncomeLess75to99_count','IncomeLess75to99_percent','Income100Plus_count',
                                'Income100Plus_percent','Median_Income_Total']]

clean_df.head()

#Start index from 1 in clean_df and rename index field
clean_df.index = np.arange(1, len(clean_df)+1)
clean_df.index = clean_df.index.set_names(['household_income_by_community_id'])

clean_df.head()

#Rename df names to match Postgres table
clean_df.rename(columns={'IncomeLess35000_count': 'incomeless35000_count'
                         ,'IncomeLess35000_percent':'incomeless35000_percent'
                        ,'IncomeLess35to49_count':'incomeless35to49_count'
                        ,'IncomeLess35to49_percent':'incomeless35to49_percent'
                        ,'IncomeLess50to74_count':'incomeless50to74_count'
                         ,'IncomeLess50to74_percent':'incomeless50to74_percent'
                        ,'IncomeLess75to99_count':'incomeless75to99_count'
                        ,'IncomeLess75to99_percent':'incomeless75to99_percent'
                        ,'Income100Plus_count':'income100plus_count'
                        ,'Income100Plus_percent':'income100plus_percent'
                        ,'Median_Income_Total':'median_income_total'}, inplace=True)
clean_df.head()

# Load household income data to Postgres table

#append data to Postgres existing table
clean_df.to_sql('household_income_by_community', engine,if_exists='append')

#Save data to a csv file
output_data_file = "target_files/MLS_Income_by_Community.csv"
output_path = os.path.join(output_data_file)
clean_df.to_csv(output_path, header=True)
