==> 6_MLS_Race_Data_Extract.ipynb <==
# ETL Project - EXTRACTION
## Terrence Cummings
Data: Race mix of Minneapolis neighborhoods scraped from mncompass.org

import pandas as pd
from bs4 import BeautifulSoup
import json
import requests
import pymongo
from splinter import Browser
from selenium import webdriver
import time
import sys
import os
import datetime
from selenium.webdriver.chrome.options import Options
chrome_options = Options()
chrome_options.add_argument("--headless")


#Initialize list to hold neighborhood URL's and neighborhood names
nbhd_links_lst = []
nbhd_names_lst = []

#URL for the MSP neighborhoods page at MN Compass. This page contains all the links to the individual neighborhood pages.
msp_nbhds_url = 'http://www.mncompass.org/profiles/neighborhoods/minneapolis-saint-paul'

#Create Chrome web driver
driver = webdriver.Chrome(options=chrome_options)

#Send the webdriver to the link
driver.get(msp_nbhds_url)

#Extract an object containing all the neighborhood links. Uses Selenium Xpath.
nbhd_links = driver.find_elements_by_xpath('//*[@id="list"]/div/ul[2]/li[*]/a')

#Go through each Selenium item and extract the link for the neighborhood.
for nbhd in nbhd_links:
    href = nbhd.get_attribute('href')
    #Add the link to the list of neighborhood links
    nbhd_links_lst.append(href)
    #Extract the name of the neighborhood from the end of the URL and add to the list of neighborhood names.
    nbhd_names_lst.append(href[60:])

#Close the webdriver    
driver.close()



nbhd_links_df = pd.DataFrame(nbhd_links_lst, columns=["URL"])
nbhd_links_df.to_csv('nbhd_links.csv', index=False)
nbhd_links_lst

#Open the webdriver
driver = webdriver.Chrome(options=chrome_options)

#Initialize lists to hold the count and percentage for each race. The list is in the same order as the neighborhoods above.
white_cnt_lst = []
white_pct_lst = []
black_cnt_lst = []
black_pct_lst = []
native_cnt_lst = []
native_pct_lst = []
asian_cnt_lst = []
asian_pct_lst = []
other_cnt_lst = []
other_pct_lst = []
twoormore_cnt_lst = []
twoormore_pct_lst = []
hispanic_cnt_lst = []
hispanic_pct_lst = []
ofcolor_cnt_lst = []
ofcolor_pct_lst = []

#For each neighborhood
for nbhd_link in nbhd_links_lst:
    #Go to the link for the neighborhood
    driver.get(nbhd_link)
    #A short timer to give the browser time to respond
    time.sleep(1)
    #Use Selenium Xpath to extract the table list for the race, access the count and percent elements of the list and add to lists
    nbhd_white = driver.find_elements_by_xpath('//*[@id="totalWhite"]')
    white_cnt_lst.append(nbhd_white[0].text)
    white_pct_lst.append(nbhd_white[2].text)
    nbhd_black = driver.find_elements_by_xpath('//*[@id="totalBlack"]')
    black_cnt_lst.append(nbhd_black[0].text)
    black_pct_lst.append(nbhd_black[2].text)
    nbhd_native = driver.find_elements_by_xpath('//*[@id="totalNative"]')
    native_cnt_lst.append(nbhd_native[0].text)
    native_pct_lst.append(nbhd_native[2].text)
    nbhd_asian = driver.find_elements_by_xpath('//*[@id="totalAsian"]')
    asian_cnt_lst.append(nbhd_asian[0].text)
    asian_pct_lst.append(nbhd_asian[2].text)
    nbhd_other = driver.find_elements_by_xpath('//*[@id="totalOther"]')
    other_cnt_lst.append(nbhd_other[0].text)
    other_pct_lst.append(nbhd_other[2].text)
    nbhd_twoormore = driver.find_elements_by_xpath('//*[@id="twoOrMoreRaces"]')
    twoormore_cnt_lst.append(nbhd_twoormore[0].text)
    twoormore_pct_lst.append(nbhd_twoormore[2].text)
    nbhd_hispanic = driver.find_elements_by_xpath('//*[@id="totalHispanic"]')
    hispanic_cnt_lst.append(nbhd_hispanic[0].text)
    hispanic_pct_lst.append(nbhd_hispanic[2].text)
    nbhd_ofcolor = driver.find_elements_by_xpath('//*[@id="totalOfColor"]')
    ofcolor_cnt_lst.append(nbhd_ofcolor[0].text)
    ofcolor_pct_lst.append(nbhd_ofcolor[2].text)

#Close the browser
driver.close()



#Zip the lists together and make a final dataframe where each row is a neighborhood.
nbhd_race_df = pd.DataFrame(list(zip(nbhd_names_lst, white_cnt_lst, white_pct_lst, black_cnt_lst, black_pct_lst, native_cnt_lst, native_pct_lst, asian_cnt_lst, asian_pct_lst, other_cnt_lst, other_pct_lst, twoormore_cnt_lst, twoormore_pct_lst, hispanic_cnt_lst, hispanic_pct_lst, ofcolor_cnt_lst, ofcolor_pct_lst, nbhd_links_lst)), columns =['neighborhood', 'white_cnt', 'white_pct', 'black_cnt', 'black_pct', 'native_cnt', 'native_pct', 'asian_cnt', 'asian_pct', 'other_cnt', 'other_pct', 'two_or_more_cnt', 'two_or_more_pct', 'hispanic_cnt', 'hispanic_pct', 'of_color_cnt', 'of_color_pct', 'URL']) 


nbhd_race_df.to_csv('nbhd_race.csv', index=False)
nbhd_race_df

#Initialize list to hold community URL's and community names
cmnty_links_lst = []
cmnty_names_lst = []

#URL for the MSP community page at MN Compass. This page contains all the links to the individual community pages.
msp_cmnty_url = 'http://www.mncompass.org/profiles/neighborhoods/minneapolis-saint-paul'

#Create Chrome web driver
driver = webdriver.Chrome(options=chrome_options)

#Send the webdriver to the link
driver.get(msp_cmnty_url)

#Extract an object containing all the commumity links. Uses Selenium Xpath.
cmnty_links = driver.find_elements_by_xpath('//*[@id="list"]/div/ul[3]/li[*]/a')

#Go through each Selenium item and extract the link for the community.
for cmnty in cmnty_links:
    href = cmnty.get_attribute('href')
    #Add the link to the list of community links
    cmnty_links_lst.append(href)
    #Extract the name of the community from the end of the URL and add to the list of community names.
    cmnty_names_lst.append(href[58:])

#Close the webdriver    
driver.close()
cmnty_links_lst

cmnty_links_df = pd.DataFrame(cmnty_links_lst, columns=["URL"])
cmnty_links_df.to_csv('cmnty_links.csv', index=False)
cmnty_links_df

#Open the webdriver
driver = webdriver.Chrome(options=chrome_options)

#Initialize lists to hold the count and percentage for each race. The list is in the same order as the neighborhoods above.
white_cnt_lst = []
white_pct_lst = []
black_cnt_lst = []
black_pct_lst = []
native_cnt_lst = []
native_pct_lst = []
asian_cnt_lst = []
asian_pct_lst = []
other_cnt_lst = []
other_pct_lst = []
twoormore_cnt_lst = []
twoormore_pct_lst = []
hispanic_cnt_lst = []
hispanic_pct_lst = []
ofcolor_cnt_lst = []
ofcolor_pct_lst = []

#For each community
for cmnty_link in cmnty_links_lst:
    #Go to the link for the neighborhood
    driver.get(cmnty_link)
    #A short timer to give the browser time to respond
    time.sleep(1)
    #Use Selenium Xpath to extract the table list for the race, access the count and percent elements of the list and add to lists
    white = driver.find_elements_by_xpath('//*[@id="totalWhite"]')
    white_cnt_lst.append(white[0].text)
    white_pct_lst.append(white[2].text)
    black = driver.find_elements_by_xpath('//*[@id="totalBlack"]')
    black_cnt_lst.append(black[0].text)
    black_pct_lst.append(black[2].text)
    native = driver.find_elements_by_xpath('//*[@id="totalNative"]')
    native_cnt_lst.append(native[0].text)
    native_pct_lst.append(native[2].text)
    asian = driver.find_elements_by_xpath('//*[@id="totalAsian"]')
    asian_cnt_lst.append(asian[0].text)
    asian_pct_lst.append(asian[2].text)
    other = driver.find_elements_by_xpath('//*[@id="totalOther"]')
    other_cnt_lst.append(other[0].text)
    other_pct_lst.append(other[2].text)
    twoormore = driver.find_elements_by_xpath('//*[@id="twoOrMoreRaces"]')
    twoormore_cnt_lst.append(twoormore[0].text)
    twoormore_pct_lst.append(twoormore[2].text)
    hispanic = driver.find_elements_by_xpath('//*[@id="totalHispanic"]')
    hispanic_cnt_lst.append(hispanic[0].text)
    hispanic_pct_lst.append(hispanic[2].text)
    ofcolor = driver.find_elements_by_xpath('//*[@id="totalOfColor"]')
    ofcolor_cnt_lst.append(ofcolor[0].text)
    ofcolor_pct_lst.append(ofcolor[2].text)

#Close the browser
driver.close()


#Zip the lists together and make a final dataframe where each row is a commumity.
cmnty_race_df = pd.DataFrame(list(zip(cmnty_names_lst, white_cnt_lst, white_pct_lst, black_cnt_lst, black_pct_lst, native_cnt_lst, native_pct_lst, asian_cnt_lst, asian_pct_lst, other_cnt_lst, other_pct_lst, twoormore_cnt_lst, twoormore_pct_lst, hispanic_cnt_lst, hispanic_pct_lst, ofcolor_cnt_lst, ofcolor_pct_lst, cmnty_links_lst)), columns =['community', 'white_cnt', 'white_pct', 'black_cnt', 'black_pct', 'native_cnt', 'native_pct', 'asian_cnt', 'asian_pct', 'other_cnt', 'other_pct', 'two_or_more_cnt', 'two_or_more_pct', 'hispanic_cnt', 'hispanic_pct', 'of_color_cnt', 'of_color_pct', 'URL']) 

cmnty_race_df.to_csv('cmnty_race.csv', index=False)
cmnty_race_df
