==> 7_Police_Incidents_ETL.ipynb <==
import pandas as pd
import datetime
import numpy as np

# Import and Read CSVs

PI_2015 = ("source_files/Police_Incidents_2015.csv")
PI_2016 = ("source_files/Police_Incidents_2016.csv")
PI_2017 = ("source_files/Police_Incidents_2017.csv")
PI_2018 = ("source_files/Police_Incidents_2018.csv")
PI_2018_2019 = ("source_files/Police_Incidents_Last_2Years.csv")
Neigh_Key = ("target_files/MLS_Neighborhoods.csv")

PI_2015_df = pd.read_csv(PI_2015)
PI_2016_df = pd.read_csv(PI_2016)
PI_2017_df = pd.read_csv(PI_2017)
PI_2018_df = pd.read_csv(PI_2018)
PI_Last2Years_df = pd.read_csv(PI_2018_2019)
Neigh_Key_df = pd.read_csv(Neigh_Key)

# Merge Years of Police Incident Data

PI_pre18 = [PI_2015_df,PI_2016_df,PI_2017_df,PI_2018_df]
PI_pre18_merge = pd.concat(PI_pre18)
PI_pre18_merge.head(1)

PI_pre18_final = PI_pre18_merge.drop(columns={"PublicAddress", "ControlNbr",
                                                "Precinct", "BeginDate", "Time",
                                                 "EnteredDate", "GBSID","X", "Y",
                                                 "LastChanged", "LastUpdateDate",
                                                 "Description","UCRCode","FID"})
PI_pre18_final.head(5)

PI_pre18_final = PI_pre18_final[["CCN","ReportedDate","Offense","Neighborhood","Lat","Long"]]
PI_pre18_final.head(5)

PI_Last2Years_df.head(2)

PI_Last2Years_final = PI_Last2Years_df.drop(columns={"publicaddress","reportedTime", "reportedDateTime",
                                                "precinct", "beginDate", "beginTime",
                                                 "enteredDate", "X", "Y",
                                                 "lastchanged", "LastUpdateDateETL","enteredDate",
                                                 "centerX","centerY","centergbsid",
                                                 "enteredDate", "description",
                                                 "OBJECTID", "UCRCode", "lastchanged","LastUpdateDateETL"})

PI_Last2Years_final.head(2)

PI_Last2Years_final = PI_Last2Years_final[['caseNumber','reportedDate','offense','neighborhood','centerLong','centerLat']]
PI_Last2Years_final.head(2)

PI_pre18_final = PI_pre18_final.rename(columns={"CCN":"caseNumber", "ReportedDate":"reportedDate","Offense":"offense","Neighborhood":"neighborhood"})
PI_Last2Years_final = PI_Last2Years_final.rename(columns={"centerLong":"Long","centerLat":"Lat"})


yearjoins = [PI_pre18_final,PI_Last2Years_final]
PI_Final = pd.concat(yearjoins)
PI_Final.head(5)

PI_Final['caseNumber'] = PI_Final['caseNumber'].str.replace(" ","")
PI_Final.head(5)

incident_id = PI_Final.index + 1

PI_Final['incident_id']= incident_id

PI_Final['neighborhood'].isnull().sum(axis = 0)

PI_Final = PI_Final.dropna(subset=['neighborhood'])

#PI_Final['neighborhood'].isnull().sum(axis = 0)

PI_Final.head(5)

# Change neighborhood names for joining with key table

PI_Final['neighborhood']= PI_Final['neighborhood'].str.upper()

PI_Final = PI_Final.replace(to_replace = "ST. ANTHONY WEST", value ="ST ANTHONY WEST")
PI_Final = PI_Final.replace(to_replace = "ST. ANTHONY EAST", value ="ST ANTHONY EAST")
PI_Final = PI_Final.replace(to_replace = "HUMBOLDT INDUSTRIAL AREA", value ="HUMBOLDT INDUST AREA")
PI_Final = PI_Final.replace(to_replace = "CAMDEN INDUSTRIAL", value ="CAMDEN INDUSTRIAL AREA")        
PI_Final = PI_Final.replace(to_replace = "CARAG", value ="South Uptown")
PI_Final = PI_Final.replace(to_replace = "BRYN - MAWR", value ="Bryn Mawr")
PI_Final = PI_Final.replace(to_replace = "COLUMBIA PARK", value ="Columbia")
PI_Final = PI_Final.replace(to_replace = "MID - CITY INDUSTRIAL", value ="Mid City Industrial Ar")
PI_Final = PI_Final.replace(to_replace = "NICOLLET ISLAND - EAST BANK", value ="Nicollet Is/East Bank")
PI_Final = PI_Final.replace(to_replace = "UNIVERSITY OF MINNESOTA", value ="University")
PI_Final = PI_Final.replace(to_replace = "STEVENS SQUARE - LORING HEIGHTS", value ="Steven's Square - Loring Heights")
PI_Final['neighborhood']= PI_Final['neighborhood'].str.upper()

PI_Final.head(5)

Neigh_Key_df['name']=Neigh_Key_df['name'].str.upper()
Neigh_Key_df = Neigh_Key_df.rename(columns={'name':"neighborhood"})
Neigh_Key_df.head(5)

PI_Final['neighborhood'].isnull().sum(axis = 0)

len(PI_Final)

# Join with key table

PI_Final_ID = pd.merge(PI_Final, Neigh_Key_df, on='neighborhood', how= 'left')
PI_Final_ID.head(5)

PI_Final_ID.dtypes

PI_Final_ID['neighborhood_id'] = PI_Final_ID['neighborhood_id'].values.astype(np.int64)
PI_Final_ID['community_id'] = PI_Final_ID['community_id'].values.astype(np.int64)
PI_Final_ID.dtypes

PI_Final_ID.head(5)

#QA Merging
#len(PI_Final_ID['neighborhood_id'])
#PI_Final_ID['neighborhood_id'].count()
#PI_Final_ID['neighborhood_id'].isnull().sum(axis = 0)

#QA Neighborhood Names
#a = PI_Final['neighborhood'].unique()
#a_df = pd.DataFrame(data=a)
#a_df 
#a_df.to_csv("target_files/neighborhoods.csv", index=False, header=True)

PI_Final_ID.head(5)

# Export CSV

PI_Final_ID.to_csv("target_files/Police_Incidents.csv", index=False, header=True)
